requirements.txt
python-telegram-bot==20.8
PyMuPDF==1.23.7
openpyxl==3.1.2
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
